feat: #1486 add fees cell from market-info in tooltip
