+++ libs/market-info/src/components/fees-breakdown/index.ts
@@ -1,2 +1,1 @@
-export * from './market-info';
 export * from './fees-breakdown';
